requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
Pillow==10.0.1
pywin32==306
ctypes 